Another variant of arithmetic decoding of the model.
The previous variant of arithmetic encoding of the model was fun.
It helped me to understand several fine points about AE.
However, from practical perspective it was too complex
and it didn't compress the model particularly well, because on the one hand
4 ranges is too little for good entropy coding, but more than for ranges
will consume too much space in the header, which is poorly amortized
by short section, that we are encoding.

So now I switched to much simpler variant. It devided the value space into
8 fixed ranges, each next range eaxctly 4 times bigger than the previous one:
1..3, 4..15, 16..63, ..., 2^14..2^16-1.
It's certainly less fun, but the code is much simpler and compression is somewhat better.