Deferred NR
Change a little the way of calculation of the range in encoder.
The cahnge makes it easier for decoder to maintain invRange <= 1/range
which in turn allows to do NR iteration of invRange every Nth iteration
instead of each end every iteration of decode loop.
In this version I am doing NR in one out of 16 iterations. It can be done
even rarer, but further reduction will have almost no effect on speed of
decoding.