16Kfp1 decode - fix bug and change the method
Don't use adjustment to invRange. So now val2c() can sometimes undershoot,
but will never overshoot.
The bug was in ignoring the case where there is one or more zero range between
result of val2c() and correct value of c.
It seems, the same bug exists both in 16Kfp2 and in 16Kz.