Faster BWT sort
New implementation is not super-fast: it is only 2-4 times faster
than previous quick and dirty variant. And it is requires more memory
for temporary buffers.
The main advantage is robustness in big-O sense.
A new algorithm is O(N*logN*logN) both on average and in worst case.
In the previous algorithm the worst case was O(N*N*logN).