bwt_sort - more improvements
1. Change layout of records of segments that requires further sorting.
Store records continuously instead of sparsely. In theory that should improve
locality of reference in wrk buffer. In parctice it has niglidgible effect.
The main purpuse of a change is preparation for sorting of index-value
pairs without indirection, where I will want continuous free space in wrk[].
2. Spicialize a common case of segment length = 2.
That change gave a decent speedup - 5 to 10%.