bwt_sort - replace quicksort with mergesort
On average "flat" mergesort have about the same speed or even a little slower
than "flat" quicksort. But the worst case is a lot better.
I thougth that the worst case for quicksort is unlikely, but after
testing with Calgary and Cantrbry test corpora it turned out that
it happens.

I still have quicksort in the code for the rare case when a work buffer is
too small for "flat" variant of algorithm. I left it, beacuse an average
speed of "indirect" mergesort is poor - 2-3 times slower than quicksort.
So, the speed is still not robust. It is still possible to construct
pathologically slow N*N*Log(N) cases. But I don't believe that they
could happen unintentionally.