32K_bwt_mtf_rle_cp - give up on separate of plains
I realized that separte encoding of 2 plains one after another is not a
very good idea. The biggest downside is on decode side - mtf decoding
can't be integrated into arithmetic decoding, but has to be done as
an additional path. That's not good for speed of decoding.
So, I am back to interleaved encoding.