32K_bwt_mtf_rle_bin - dynamic encoding for non-MS bits of zrle
Change handling of non-MS bits of zero run length.
Treat both midT==0 and midT==1 as counters that are separate from
the same positions for MS bits of the run. Don't assume, as in the
previous version, equal probability for RUNA and RUNB symbols.
An improvment of compression ratio is tiny, but I like consistency
of the new variant.