32K_bwt_mtf_rle_bbin - done
As expected, expansion is ~1.2x faster than 32K_bwt_mtf_rle_bin variants
while compression ratio and compression speed is approximately the same.
So, among bi-section based attempts that variant is so far the best.

Still, both compression and expansion is 1.5x to 2.5x slower than
32K_bwt_mtf_rle_cb and compression ratio is not significantly better.
So, I have to conclude that the whole idea of bi-section, while elegant,
does not appear to be practical.