32K_bwt_mtf_rle_bin - algorithm change
In case when statistics for a current chunk are close to those of the previous
chunk, use exact (i.e. not quantized) statistics of the previous chunk
to chose a range of arithmetic encoder.
The advantage is not a denser encoding of the bulk of the data (in fact
it's a little less dense), but increased amount of 'qh is the same' codes
in the model = shorter model.
But overall improvement of compression ratio is very small.